% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abbreviations}
\alias{abbreviations}
\title{Common abbreviations to be substituted}
\format{
\subsection{\code{abbreviations}}{

A data frame with 2 columns:
\describe{
\item{regex}{regex to be replaced}
\item{replacement}{replacement}
}
}
}
\usage{
abbreviations
}
\description{
Common abbreviations to be substituted
}
\keyword{datasets}
