% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{allgrams_tibble}
\alias{allgrams_tibble}
\title{returns all unigrams to ngram_max tokens from a given phrase plus their
phonetic encoding and some statistics}
\usage{
allgrams_tibble(x, ngram_max = 3, mcl = 10, modify_words = NULL)
}
\arguments{
\item{x}{character phrase to be encoded}

\item{ngram_max}{the longest ngram to be extracted}
}
\value{
tibble with columns token, phon, n, nword, nstop
}
\description{
returns all unigrams to ngram_max tokens from a given phrase plus their
phonetic encoding and some statistics
}
\details{
for the phrase "the seagull is dead" with ngram_max = 2 the tokens
returned would be
the
seagull
is
dead
the seagull
seagull is
is dead
}
