% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{assign_home}
\alias{assign_home}
\title{determines whether or not an assault location description should be coded
to the individuals home address}
\usage{
assign_home(txt, col = ald_clean, w2v_file = "w2v_model_cbow_8_300.bin")
}
\arguments{
\item{txt}{data frame with a column containing the text to be assessed}

\item{col}{name of the column containing text to be assessed}
}
\value{
vector of the same length as txt containing "home" or "other"
}
\description{
determines whether or not an assault location description should be coded
to the individuals home address
}
