% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{assign_home_simple}
\alias{assign_home_simple}
\title{determines whether or not an assault location description should be coded
to the individuals home address - uses a very simple process of matching a
few key phrases without relying on any externally trained model}
\usage{
assign_home_simple(txt, col = ald_clean)
}
\arguments{
\item{txt}{data frame with a column containing the text to be assessed}

\item{col}{name of the column containing text to be assessed}
}
\value{
vector of the same length as txt containing "home" or "other"
}
\description{
determines whether or not an assault location description should be coded
to the individuals home address - uses a very simple process of matching a
few key phrases without relying on any externally trained model
}
