% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{clean_text}
\alias{clean_text}
\title{standard text cleaning}
\usage{
clean_text(d, col = assault_location_description, abbreviations = NULL)
}
\arguments{
\item{d}{data frame containing text to be cleaned}

\item{col}{column name returning text to be cleaned}
}
\value{
d data frame with cleaned text in column ald_clean
}
\description{
standard text cleaning
}
\details{
text cleaning removes common abbreviations, punctuation
}
