% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{countries with values to be appended to a place name data frame}
\format{
\subsection{\code{countries}}{

A data frame with ... rows and 6 columns:
\describe{
\item{name}{name of country}
\item{easting, northing}{position on BNG, NA in this case}
\item{id}{unique ID}
\item{class}{country in this case}
}
}
}
\source{
\url{https://www.data.gov.uk/dataset/0b412b57-6934-4f28-b495-640cdc7e8f7f/foreign-commonwealth-and-development-office-geographical-names-index/datafile/f9835358-6435-41da-803c-f7917db7fdad/preview}
}
\usage{
countries
}
\description{
countries with values to be appended to a place name data frame
}
\keyword{datasets}
