% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode.R
\name{detect_outcode}
\alias{detect_outcode}
\title{Detects whether or not a string x contains one of the outcodes listed in
the vector s}
\usage{
detect_outcode(x, s)
}
\arguments{
\item{x}{the string to be tested}

\item{s}{a vector of valid outcodes for the relevant area}
}
\value{
logical indicating the presence of a valid outcode in x
}
\description{
An outcode (also known as postcode district) is the part of a UK postcode
before the space, e.g. SW19, EC1A
}
\details{
this could be a simple regex, but by comparing against legitimate
outcodes properly formatted but invalid outcodes are rejected
}
\examples{
utils::data(pndf)
pcd <- pndf |>
  dplyr::filter(class == "pcd") |>
  dplyr::select(name) |>
  unlist(use.names = TRUE)
detect_outcode("the outcode is sw1a in this example", pcd)

}
