% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docs_fit}
\alias{docs_fit}
\title{model for assigning home locations}
\format{
\subsection{\code{docs_fit}}{

A data frame with ... rows and 6 columns:
}
}
\usage{
docs_fit
}
\description{
model for assigning home locations
}
\keyword{datasets}
