% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Calculates distance between two points}
\usage{
euclidean_dist(x1, y1, x2, y2)
}
\arguments{
\item{x1, y1, x2, y2}{xy coordinates of the two points (normally 6 digit BNG)}
}
\value{
distance in coordinate units (normally m)
}
\description{
Calculates distance between two points
}
