% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode.R
\name{extract_whole_postcode}
\alias{extract_whole_postcode}
\title{given an outcode extracts the whole or just the postcode sector from x
using an existing outcode seems to reduce the possibility of extracting
the alphanumeric typos often present in clincial notes}
\usage{
extract_whole_postcode(x, outcode, sector = FALSE)
}
\arguments{
\item{x}{character containing a postcode}

\item{outcode}{character outcode}

\item{sector}{boolean look for a whole postcode or just postcode sector}
}
\value{
character postcode (lower case)
}
\description{
given an outcode extracts the whole or just the postcode sector from x
using an existing outcode seems to reduce the possibility of extracting
the alphanumeric typos often present in clincial notes
}
\details{
outcode can be extracted from the pndf data frame
}
\examples{
extract_whole_postcode("the postcode is sw1a 1aa in this example", "sw1a")

}
