% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_str_dup}
\alias{is_str_dup}
\title{checks if a string is duplicated internally}
\usage{
is_str_dup(x)
}
\arguments{
\item{x}{character which may have duplication}
}
\value{
deduplicated character
}
\description{
checks if a string is duplicated internally
}
\details{
if a character contains something like "this word this word" will
de-duplicate and return "this word". Duplication seems to be quite a common
feature of
}
