% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{match_place}
\alias{match_place}
\title{match x against the closest match in a place name data frame with columns
id and name}
\usage{
match_place(x, pndf, mcl = 10, precalc_tokens = NULL, ngram_max = 3)
}
\arguments{
\item{x}{character a short sentence such as that recorded in the assault
location description field}

\item{pndf}{data frame a place name data frame such as that supplied by istv::pndf}

\item{mcl}{integer for phonetic encoding as documented in phonics::refinedSoundex}

\item{precalc_tokens}{data frame if supplied contains previously calculated tokens
and phonetic encoding for pndf. If NULL will calculate.}
}
\value{
all possible candidate matches
}
\description{
match x against the closest match in a place name data frame with columns
id and name
}
\details{
supplying precalc_tokens will make this quite a bit quicker
}
