% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max0}
\alias{max0}
\title{returns max(x) from a vector if it has rows, if it has no rows returns
0}
\usage{
max0(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric
}
\description{
returns max(x) from a vector if it has rows, if it has no rows returns
0
}
