% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{ngram_check}
\alias{ngram_check}
\title{checks if a string contains a phrase from a vector, returns the number of
words in that phrase}
\usage{
ngram_check(x, check_words = istv::place_names)
}
\arguments{
\item{x}{character to be checked}

\item{check_words}{character vector containing dictionary word place names}
}
\value{
integer count of the number of words
}
\description{
useful to modify the number of dictionary words for names like "mile end"
(both of which are dictionary words but together indicate a place)
}
\details{
add
}
