% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{ngram_tibble}
\alias{ngram_tibble}
\title{a windowing function that sweeps through a unigram tibble and
extracts ngrams}
\usage{
ngram_tibble(x, ngram = 2, modify_words = NULL)
}
\arguments{
\item{x}{data frame the unigram tibble}

\item{ngram}{integer ngram, e.g. 2 for bigrams}
}
\value{
tibble with columns token, phon, n, nword, nstop
}
\description{
a windowing function that sweeps through a unigram tibble and
extracts ngrams
}
\details{
the returned tibble includes columns n the ngram of the token, nword
the number of dictionary words in the token and nstop the number of stopwords
in the token (see word_check() and stopword_check()) for the definitions
}
