% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{ngram_window}
\alias{ngram_window}
\title{helper function for ngram_tibble}
\usage{
ngram_window(x, ngram)
}
\arguments{
\item{x}{data frame the unigram tibble}

\item{ngram}{integer ngram, e.g. 2 for bigrams}
}
\value{
tibble with columns token, phon, n, nword, nstop
}
\description{
helper function for ngram_tibble
}
\details{
the returned tibble includes columns n the ngram of the token, nword
the number of dictionary words in the token and nstop the number of stopwords
in the token (see word_check() and stopword_check()) for the definitions
}
