% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openname_headers}
\alias{openname_headers}
\title{headers of openname file from ODS}
\format{
\subsection{\code{openname_headers}}{

Character vector of openname column headers
}
}
\source{
Ordnance Survey documentation
}
\usage{
openname_headers
}
\description{
headers of openname file from ODS
}
\keyword{datasets}
