% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{place_names}
\alias{place_names}
\title{place names that are also dictionary words, e.g. derby}
\format{
\subsection{\code{place_names}}{

Character vector of place names
}
}
\source{
Ordnance Survey documentation
}
\usage{
place_names
}
\description{
place names that are also dictionary words, e.g. derby
}
\keyword{datasets}
