% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pndf}
\alias{pndf}
\title{place name dataframe}
\format{
\subsection{\code{pndf}}{

A data frame with ... rows and 6 columns:
\describe{
\item{name}{name}
\item{easting, northing}{BNG coordinates of location}
\item{id}{unique id}
\item{class}{class - e.g. town or street}
\item{id_parent}{id of parent entity if applicable}
}
}
}
\source{
\url{https://www.odsdatasearchandexport.nhs.uk/}
}
\usage{
pndf
}
\description{
place name dataframe
}
\keyword{datasets}
