% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precalc_tokens}
\alias{precalc_tokens}
\title{place name tokens dataframe}
\format{
\subsection{\code{precalc_tokens}}{

A data frame with ... rows and 6 columns:
\describe{
\item{name}{id matching to pndf}
\item{token}{word or ngram}
\item{phon}{phonetic encoding of the token}
\item{n}{ngram of the token}
\item{nstop}{number of stopwords in the token}
\item{nword}{number of dictionary words in the token}
}
}
}
\source{
\url{https://www.odsdatasearchandexport.nhs.uk/}
}
\usage{
precalc_tokens
}
\description{
place name tokens dataframe
}
\keyword{datasets}
