% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pndf.R
\name{precalculate_pndf_tokens}
\alias{precalculate_pndf_tokens}
\title{precalculates word and phonetic tokens for a given place name data frame
this will take hours on most desktop machines but once done substantially
speeds up geocoding}
\usage{
precalculate_pndf_tokens(pndf, ngram_max = 2, write = TRUE)
}
\arguments{
\item{pndf}{data frame with (at least) columns id and name}

\item{ngram_max}{integer the longest ngram to calculate}
}
\value{
data frame invisible, with columns id, token, phon, n, nword, nstop
as documented in match_place function
}
\description{
precalculates word and phonetic tokens for a given place name data frame
this will take hours on most desktop machines but once done substantially
speeds up geocoding
}
\details{
saves the output data frame to the package object pndf_tokens,
overwrites previous versions
}
