% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_xy}
\alias{site_xy}
\title{London Type 1 (consultant led) Emergency Department site locations}
\format{
\subsection{\code{sites_xy}}{

A data frame with 28 rows and 7 columns:
\describe{
\item{icb_name, trust_name, site_name}{human readable names for icb, trust and site}
\item{icb, trust, site}{ODS codes for icb, trust and site}
\item{geometry}{Coordinates of the site on BNG}
}
}
}
\source{
\url{https://www.odsdatasearchandexport.nhs.uk/}
}
\usage{
site_xy
}
\description{
Coordinates of London Type 1 Emergency Departments, their ODS codes and
associated trust and ICB. EPSG:27700 (British National Grid)
}
\keyword{datasets}
