% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{stopword_check}
\alias{stopword_check}
\title{checks if x is a stopword}
\usage{
stopword_check(x, lexicon = "onix")
}
\arguments{
\item{x}{a character, preferably a single word}

\item{lexicon}{a character as documented in tidytext::stop_words}
}
\value{
boolean
}
\description{
checks if x is a stopword
}
