% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{str_dedup}
\alias{str_dedup}
\title{checks if a string is duplicated internally and returns the deduplicated string}
\usage{
str_dedup(x)
}
\arguments{
\item{x}{character which may have duplication}
}
\value{
deduplicated character
}
\description{
checks if a string is duplicated internally and returns the deduplicated string
}
\details{
if a character contains something like "this word this word" will
de-duplicate and return "this word". Duplication seems to be quite a common
feature of
}
