% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{unigram_tibble}
\alias{unigram_tibble}
\title{makes a tibble containing all unigrams and their phonetic mapping based on
refined soundex algorithm for any given string}
\usage{
unigram_tibble(x, mcl = 10)
}
\arguments{
\item{x}{character a string containing the phrase to be encoded}

\item{mcl}{integer maximum code length for refinedSoundex}
}
\value{
tibble with columns token, phon, n, nword, nstop
}
\description{
makes a tibble containing all unigrams and their phonetic mapping based on
refined soundex algorithm for any given string
}
\details{
the returned tibble includes columns n (1, indicating unigram), nword
the number of dictionary words in the token and nstop the number of stopwords
in the token (see word_check() and stopword_check()) for the definitions
}
