% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{word_check}
\alias{word_check}
\title{checks whether a word is a non-dictionary word, where the definition of
non-dictionary includes the names of larger populated places, e.g.
Barking.}
\usage{
word_check(x, place_names = NULL)
}
\arguments{
\item{x}{character to be checked}

\item{place_names}{vector of place names that are the same as common words}
}
\value{
boolean whether x is a word
}
\description{
checks whether a word is a non-dictionary word, where the definition of
non-dictionary includes the names of larger populated places, e.g.
Barking.
}
\details{
note that the definition of words will exclude proper nouns so most
place names (e.g. London) will not count as words. Some place names (e.g.
Barking) are polysemic which is why we need the place_names vector
}
