% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_outcode}
\alias{extract_outcode}
\title{given s, a vector of valid outcodes, extracts the first detected present
in x}
\usage{
extract_outcode(x, s, fil = TRUE)
}
\arguments{
\item{x}{string with a postcode district to be extracted}

\item{s}{vector of valid postcode districts}

\item{fil}{remove punctuation in x}
}
\value{
character outcode (lower case)
}
\description{
given s, a vector of valid outcodes, extracts the first detected present
in x
}
\details{
it seems to be faster to use detect than extract rather than
attempting to extract on x
}
\examples{
utils::data(pndf)
pcd <- pndf |>
  dplyr::filter(class == "pcd") |>
  dplyr::select(name) |>
  unlist(use.names = TRUE)
extract_outcode("the outcode is sw1a in this example", pcd)

}
