% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_pndf}
\alias{filter_pndf}
\title{filter a place name file to be within a certain distance of a point}
\usage{
filter_pndf(x = 531370, y = 180220, r = 500, pndf = NULL)
}
\arguments{
\item{x}{numeric, x coordinate of center (probably best on BNG)}

\item{y}{numeric, y coordinate of center (probably best on BNG)}

\item{r}{distance of bounding box from center}

\item{pndf}{tibble with columns at least easting and northing containing
coordinates of locations to be filtered, if null the data supplied with the
package is used}
}
\value{
data frame pndf filtered to area
}
\description{
filter a place name file to be within a certain distance of a point
}
\details{
creates a bounding box from x-r to x+r and y-r to y+r and filters
locations supplied in pndf
}
