% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geocode_text}
\alias{geocode_text}
\title{geocodes a data frame of text}
\usage{
geocode_text(
  txt,
  pndf = NULL,
  site_xy = NULL,
  home_fn = assign_home,
  debug = FALSE
)
}
\arguments{
\item{txt}{data frame containing columns site_code, assault_location_description plus anything else that is useful}

\item{pndf}{data frame with columns id, name, easting, northing, class, fulladdress, LSOA21CD, if NULL uses the package default}

\item{site_xy}{data frame with site_code, easting northing of the hospital sites, if NULL uses the package default}

\item{home_fn}{function to be used to assign home, returns vector of "home"/"other" values}

\item{debug}{boolean, if TRUE the output is much more voluminous and can be used to troubleshoot matching issues}
}
\value{
txt with additional columns ...
}
\description{
geocodes a data frame of text
}
\details{
extracts all the postcode variants (full, sector, district), flags if home
}
\examples{
# pndf = NULL
# site_xy = NULL
# txt <- tibble::tibble(site_code = c("RJ122", "RJ122", "RVR50", "R1H12", "RJ701", "RVR50", "R1H12", "R1H12", "RAL01", "RAL01"),
#                       mechanism = "blunt injury",
#                       injury_datetime = as_datetime("1999-12-31 23:59:59"),
#                       assault_location_description = c("- -", "London Eye, SE1", "Nandos, Epsom, KT19 8AS", "Whitechapel, E1", "sainsbury on wandsworth road sw8", "nandos in epsom town centre near the station, KT19", "on holiday in spain", "na", "on the weekend at home", "ASDA in roehampton"))
#
# g <- geocode_text(txt)

}
