% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_candidate_df}
\alias{make_candidate_df}
\title{makes a minimum spec candidate data frame with zero rows}
\usage{
make_candidate_df()
}
\value{
zero row data frame with columns id, pc_match
}
\description{
makes a minimum spec candidate data frame with zero rows
}
