% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_pndf}
\alias{make_pndf}
\title{Creates a gazetteer of place names from several sources.}
\usage{
make_pndf(
  paf_file = NULL,
  openname_dir = NULL,
  openname_headers = NULL,
  openname_types = NULL,
  greenspace_files = NULL,
  include_london_sites = TRUE,
  lad_file = NULL,
  postcode_poly_dir = NULL,
  lsoa_file = NULL,
  add_full_address = TRUE,
  write_package_data = FALSE,
  pndf_levels = c("pc", "org", "pc2", "street", "locality", "pcd", "lad",
    "populatedplace", "greenspace", "town", "landform", "country")
)
}
\arguments{
\item{paf_file}{the build environment file from PAF (may be subset)}

\item{openname_dir}{a directory containing OS open name files}

\item{openname_headers}{headers for the columns, defaults to the values current
when this file was created in March 2025}

\item{openname_types}{types to extract from the openname file - useful if
access to paf file is limited and you want to include}

\item{greenspace_files}{a vector of greenspace files}

\item{include_london_sites}{boolean whether or not to roll the London UEC
sites in}

\item{lad_file}{csv file containing LAD names and codes (LAD24CD, LAD24NM)}

\item{postcode_poly_dir}{directory containing postcode polygons}

\item{lsoa_file}{file containing LSOA polygons (preferably full resolution)}

\item{add_full_address}{boolean, add a fulladdress column to the pndf output}

\item{write_package_data}{if true write data to package data folder}

\item{pndf_levels}{vector}
}
\value{
a dataframe with columns
id
name
id_parent (may be NA)
easting on the BNG
northing
class
fulladdress
LSOA21CD
pcd
}
\description{
This function will likely take several minutes to run
}
\details{
a place name data frame can be created using one or more of these
inputs.

Sources:
\itemize{
\item postcode address file - detailed and contains venue and business names
Obtained from Ordnance Survey via the public sector geospatial agreement
recommend restricting to the most relevant area (in the case of the pndf
file supplied in the package it is a 10km buffer around London). There are
several files in the extract, and the relevant one is add_gb_builtaddress.csv
\item OS Open Name - higher level but even if PAF is in use, good to add in
towns outside of the area covered by your PAF extract. May be a substitute
for PAF if that is not available. Currently at
https://www.ordnancesurvey.co.uk/products/os-open-names
\item OS Greenspace - contains names of open spaces. Currently at
https://www.ordnancesurvey.co.uk/products/os-open-greenspace
\item LADs contains names of local authority districts - in some cases these
aren't the same as names of populated places but often used as reference.
Obtained from https://geoportal.statistics.gov.uk/, and the file to use
is the 2024 version (or field names will need to be changed, see code
below for details)
\item FCDO country names - to cover places outside the UK (this is present as a dataset in the package)
FCDO country names https://www.gov.uk/government/publications/geographical-names-and-information
Created by countries <- readr::read_csv("file") \%>\%
dplyr::mutate(id = paste0("fcdo_", \verb{Country code}), id_parent = NA, class = "country", easting = NA, northing = NA) \%>\%
dplyr::rename(name = Name) \%>\%
dplyr::select(id, name, easting, northing, id_parent, class)
\item postcode polygons - to assign postcode districts to points where that has
not already been done in the source. Obtained from OS via PSGA, it comes as
as series of shapefiles for each postcode area
\item lsoa_file - to assign all points to an LSOA for later processing
}
}
