% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{match_place}
\alias{match_place}
\title{match x against the closest match in a place name data frame with columns
id and name}
\usage{
match_place(x, pndf, mcl = 10)
}
\arguments{
\item{x}{character a short sentence such as that recorded in the assault
location description field}

\item{pndf}{data frame a place name data frame such as that supplied by istv::pndf}

\item{mcl}{integer for phonetic encoding as documented in phonics::refinedSoundex}
}
\value{
all possible candidate matches
}
\description{
match x against the closest match in a place name data frame with columns
id and name
}
