% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_abbreviations}
\alias{replace_abbreviations}
\title{replaces abbreviations using a table of regex and the replacement}
\usage{
replace_abbreviations(x, abbreviations = NULL)
}
\arguments{
\item{x}{character containing abbreviations to be replaced}

\item{abbreviations}{tibble with columns regex and replacement giving regex
for abbreviations and the relevant replacement, if none is supplied
utils::data((abbreviations) is used)}
}
\value{
character containing x with abbreviations replaced
}
\description{
replaces abbreviations using a table of regex and the replacement
}
