% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_candidate}
\alias{select_candidate}
\title{selects the best possible candidate on a candidate data frame based on several criteria}
\usage{
select_candidate(
  candidates,
  pc_m = 0.5,
  xy = c(531370, 180220),
  reject_pcd = TRUE
)
}
\arguments{
\item{candidates}{data frame such as that returned by select_candidate}

\item{pc_m}{double}

\item{xy}{integer vector of length 2, xy coordinates of site attended on BNG}

\item{reject_pcd}{boolean whether or not to reject postcode matches}
}
\value{
all possible candidate matches
}
\description{
selects the best possible candidate on a candidate data frame based on several criteria
}
\details{
some details to be filled in
}
