% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{word_check}
\alias{word_check}
\title{checks whether a word is a non-dictionary word, where the definition of
non-dictionary includes the names of larger populated places, e.g.
Barking.}
\usage{
word_check(x, place_names = NULL)
}
\arguments{
\item{x}{character to be checked}

\item{place_names}{vector of place names that are the same as common words}
}
\value{
boolean whether x is a word
}
\description{
checks whether a word is a non-dictionary word, where the definition of
non-dictionary includes the names of larger populated places, e.g.
Barking.
}
